requirements.txt
setuptools
pyinstaller
tinyaes
pywin32
requests
keyboard
psutil
flask
waitress
pillow
pyqt5
PyQtWebEngine
screeninfo
bs4